+++ Packs/GenericSQL/Integrations/GenericSQL/GenericSQL.py
@@ -5,6 +5,7 @@ from CommonServerUserPython import *
 from typing import Any, Tuple, Dict, List, Callable
 import sqlalchemy
 import pymysql
+import traceback
 from sqlalchemy.sql import text
 
 # In order to use and convert from pymysql to MySQL this line is necessary
@@ -18,7 +19,7 @@ class Client:
     """
 
     def __init__(self, dialect: str, host: str, username: str, password: str, port: str,
-                 database: str, connect_parameters: str):
+                 database: str, connect_parameters: str, ssl_connect: bool):
         self.dialect = dialect
         self.host = host
         self.username = username
@@ -26,6 +27,7 @@ class Client:
         self.port = port
         self.dbname = database
         self.connect_parameters = connect_parameters
+        self.ssl_connect = ssl_connect
         self.connection = self._create_engine_and_connect()
 
     @staticmethod
@@ -55,11 +57,19 @@ class Client:
         try:
             module = self._convert_dialect_to_module(self.dialect)
             db_preferences = f'{module}://{self.username}:{self.password}@{self.host}:{self.port}/{self.dbname}'
+            ssl_connection = {}
             if self.dialect == "Microsoft SQL Server":
                 db_preferences += "?driver=FreeTDS"
-            if self.connect_parameters:
+            if self.connect_parameters and self.dialect == "Microsoft SQL Server":
+                db_preferences += f'&{self.connect_parameters}'
+            elif self.connect_parameters and self.dialect != "Microsoft SQL Server":
+                # a "?" was already added when the driver was defined
                 db_preferences += f'?{self.connect_parameters}'
-            return sqlalchemy.create_engine(db_preferences).connect()
+
+            if self.ssl_connect:
+                ssl_connection = {'ssl': {'ssl-mode': 'preferred'}}
+
+            return sqlalchemy.create_engine(db_preferences, connect_args=ssl_connection).connect()
         except Exception as err:
             raise Exception(err)
 
@@ -74,7 +84,10 @@ class Client:
 
         result = self.connection.execute(sql_query, bind_vars)
         results = result.fetchall()
-        headers = results[0].keys()
+        headers = []
+        if results:
+            # if the table isn't empty
+            headers = results[0].keys()
         return results, headers
 
 
@@ -141,8 +154,12 @@ def sql_query_execute(client: Client, args: dict, *_) -> Tuple[str, Dict[str, An
         bind_variables = generate_bind_vars(bind_variables_names, bind_variables_values)
 
         result, headers = client.sql_query_execute_request(sql_query, bind_variables)
-        table = [dict(row) for row in result]
-        human_readable = tableToMarkdown(name="Query result:", t=table[skip:skip + limit], headers=headers,
+        # converting an sqlalchemy object to a table
+        converted_table = [dict(row) for row in result]
+        # converting b'' and datetime objects to readable ones
+        table = [{str(key): str(value) for key, value in dictionary.items()} for dictionary in converted_table]
+        table = table[skip:skip + limit]
+        human_readable = tableToMarkdown(name="Query result:", t=table, headers=headers,
                                          removeNull=True)
         context = {
             'Result': table,
@@ -171,12 +188,13 @@ def main():
     password = params.get("credentials").get("password")
     host = params.get('host')
     database = params.get('dbname')
+    ssl_connect = params.get('ssl_connect')
     connect_parameters = params.get('connect_parameters')
     try:
         command = demisto.command()
         LOG(f'Command being called in SQL is: {command}')
         client = Client(dialect=dialect, host=host, username=user, password=password,
-                        port=port, database=database, connect_parameters=connect_parameters)
+                        port=port, database=database, connect_parameters=connect_parameters, ssl_connect=ssl_connect)
         commands: Dict[str, Callable[[Client, Dict[str, str], str], Tuple[str, Dict[Any, Any], List[Any]]]] = {
             'test-module': test_module,
             'query': sql_query_execute,
@@ -188,7 +206,7 @@ def main():
             raise NotImplementedError(f'{command} is not an existing Generic SQL command')
         client.connection.close()
     except Exception as err:
-        return_error(err)
+        return_error(f'Unexpected error: {str(err)} \nquery: {demisto.args().get("query")} \n{traceback.format_exc()}')
 
 
 if __name__ in ('__main__', '__builtin__', 'builtins'):