+++ Packs/FeedOffice365/Integrations/FeedOffice365/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## [Unreleased]
-Set **bypass exclusion list** parameter to true by default.
+Set the default value for the *bypass exclusion list* parameter to "true".
 
 ## [20.3.4] - 2020-03-30
 -