+++ Packs/FeedOffice365/Integrations/FeedOffice365/CHANGELOG.md
@@ -1,6 +1,9 @@
 ## [Unreleased]
 
 
+## [20.5.0] - 2020-05-12
+Set the default value for the *bypass exclusion list* parameter to "true".
+
 ## [20.3.4] - 2020-03-30
 -